move taper to its own module
