+++ pycbc/waveform/__init__.py
@@ -1,5 +1,6 @@
 from pycbc.waveform.waveform import *
 from pycbc.waveform.utils import *
+from pycbc.waveform.taper import *
 from pycbc.waveform.bank import *
 from pycbc.waveform.generator import *
 from pycbc.waveform.ringdown import *
